requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Email Scraper/Email Scraper.ipynb: 7
# Email Scraper/email_scraper.py: 7
pandas == 1.1.2

# Email Scraper/Email Scraper.ipynb: 6
# Email Scraper/email_scraper.py: 6
python_dateutil == 2.8.1
